Remove TODO remove utf8. Keep jsonization to utf8
Rationale for this decision: JSON is by default UTF-8 encoded. We even have our header Content-Type as 'application/json; charset=utf8'. When the message body is a string, this is not a problem because Perl string is Unicode, not UTF-8. But if message body is (a hash) serialized into JSON, this becomes a problem, because the serialization (jsonization) already converts all the strings into UTF-8. So we must remind user to do jsonization without UTF-8-conversion. We shall build a quick send function that serialized any hash in the proper way. We cannot detect UTF-8 in the body scalar because detection is unreliable.

Signed-off-by: Mikko Koivunalho <mikko.koivunalho@iki.fi>