Keep using JSON::MaybeXS.
JSON::MaybeXS uses backend Cpanel::JSON::XS, with a fallback to JSON::XS and JSON::PP. I wanted to switch to JSON.pm because it seemed to matured. But seems JSON::MaybeXS has also matured and is a good alternative to JSON. I hope this is the last time I will ponder this.
Signed-off-by: Mikko Koivunalho <mikko.koivunalho@iki.fi>